requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cycler==0.10.0
Cython==0.29.23
Flask==2.0.1
gensim==4.1.2
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
nltk==3.6.3
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.2
python-firebase @ git+https://github.com/ozgur/python-firebase@0d79d7609844569ea1cec4ac71cb9038e834c355
pytz==2021.1
regex==2021.9.24
requests==2.26.0
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
smart-open==5.2.1
threadpoolctl==2.2.0
tqdm==4.62.3
urllib3==1.26.7
Werkzeug==2.0.1
